requirements.txt
streamlit==1.28.0
pandas==2.1.0
numpy==1.24.3
python-docx==1.1.0
PyPDF2==3.0.1
spacy==3.7.2
textblob==0.17.1
scikit-learn==1.3.0
nltk==3.8.1
plotly==5.17.0
pillow==10.0.0
transformers==4.35.0
torch==2.1.0
sentence-transformers==2.2.2
openai==1.3.0
python-dotenv==1.0.0
